requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click-plugins==1.1.1
click==7.0
colorama==0.4.3
idna==2.8
markovify==0.8.0
multidict==4.7.4
requests==2.22.0
shodan==1.21.3
slackclient==2.5.0
unidecode==1.1.1
urllib3==1.25.8
xlsxwriter==1.2.7
yarl==1.4.2
